fix: corregir errores de compilacion en game/
- game/CMakeLists.txt: corregir rutas duplicadas (game/game/...) y comentar archivos que no existen aun
- character.h: quitar '=' extra en declaracion de getNombre()
- charClase.h: corregir nombre de clase 'charClase' -> 'CharClase'
- charClase.cpp: corregir include 'charClases.h' -> 'charClase.h'
- enano.cpp: agregar 'const' a metodos y renombrar getRaza -> getNombre
- gnomo.cpp: agregar 'const' a metodos
- gnomo.h: corregir tipo de retorno getFRecuperacion (int -> float)
- jugador.h: corregir rutas de includes para razas y clases
- jugador.cpp: renombrar recuperacionPeriodica -> recuperacionPasiva
- game.cpp: quitar '=' extra, usar nombres correctos de metodos de Mapa
- map.h: corregir 'enum clase' -> 'enum class', agregar ';' al final de clase

+++ game/razas/gnomo.h
@@ -6,5 +6,5 @@ public:
     std::string getNombre() const override;
     int getMaxVida() const override;
     int getMaxMana() const override;
-    int getFRecuperacion() const override;
+    float getFRecuperacion() const override;
 };
